refactor: isolate Slider stories
