+++ packages/react-aria-components/stories/index.stories.tsx
@@ -11,7 +11,7 @@
  */
 
 import {action} from '@storybook/addon-actions';
-import {Button, Calendar, CalendarCell, CalendarGrid, Cell, Checkbox, Column, ColumnResizer, DateField, DateInput, DatePicker, DateRangePicker, DateSegment, Dialog, DialogTrigger, DropZone, FileTrigger, Group, Header, Heading, Input, Keyboard, Label, Link, ListBox, ListBoxItem, ListBoxProps, Menu, MenuItem, MenuTrigger, Modal, ModalOverlay, NumberField, OverlayArrow, Popover, Radio, RadioGroup, RangeCalendar, ResizableTableContainer, Row, SearchField, Section, Select, SelectValue, Separator, Slider, SliderOutput, SliderThumb, SliderTrack, Switch, Tab, Table, TableBody, TableHeader, TabList, TabPanel, Tabs, TabsProps, Tag, TagGroup, TagList, Text, TextField, TimeField, ToggleButton, Toolbar, Tooltip, TooltipTrigger, useDragAndDrop} from 'react-aria-components';
+import {Button, Calendar, CalendarCell, CalendarGrid, Cell, Checkbox, Column, ColumnResizer, DateField, DateInput, DatePicker, DateRangePicker, DateSegment, Dialog, DialogTrigger, DropZone, FileTrigger, Group, Header, Heading, Input, Keyboard, Label, Link, ListBox, ListBoxItem, ListBoxProps, Menu, MenuItem, MenuTrigger, Modal, ModalOverlay, NumberField, OverlayArrow, Popover, Radio, RadioGroup, RangeCalendar, ResizableTableContainer, Row, SearchField, Section, Select, SelectValue, Separator, Switch, Tab, Table, TableBody, TableHeader, TabList, TabPanel, Tabs, TabsProps, Tag, TagGroup, TagList, Text, TextField, TimeField, ToggleButton, Toolbar, Tooltip, TooltipTrigger, useDragAndDrop} from 'react-aria-components';
 import {classNames} from '@react-spectrum/utils';
 import clsx from 'clsx';
 import {FocusRing, isTextDropItem, mergeProps, useButton, useClipboard, useDrag} from 'react-aria';
@@ -369,76 +369,6 @@ export const DateRangePickerExample = () => (
     </Popover>
   </DateRangePicker>
 );
-
-export const SliderExample = () => (
-  <Slider
-    data-testid="slider-example"
-    defaultValue={[30, 60]}
-    style={{
-      position: 'relative',
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-      width: 300
-    }}>
-    <div style={{display: 'flex', alignSelf: 'stretch'}}>
-      <Label>Test</Label>
-      <SliderOutput style={{flex: '1 0 auto', textAlign: 'end'}}>
-        {({state}) => `${state.getThumbValueLabel(0)} - ${state.getThumbValueLabel(1)}`}
-      </SliderOutput>
-    </div>
-    <SliderTrack
-      style={{
-        position: 'relative',
-        height: 30,
-        width: '100%'
-      }}>
-      <div
-        style={{
-          position: 'absolute',
-          backgroundColor: 'gray',
-          height: 3,
-          top: 13,
-          width: '100%'
-        }} />
-      <CustomThumb index={0}>
-        <Label>A</Label>
-      </CustomThumb>
-      <CustomThumb index={1}>
-        <Label>B</Label>
-      </CustomThumb>
-    </SliderTrack>
-  </Slider>
-);
-
-export const SliderCSS = (props) => (
-  <Slider {...props} defaultValue={30} className={styles.slider}>
-    <div className={styles.label}>
-      <Label>Test</Label>
-      <SliderOutput />
-    </div>
-    <SliderTrack className={styles.track}>
-      <SliderThumb className={styles.thumb} />
-    </SliderTrack>
-  </Slider>
-);
-
-SliderCSS.args = {
-  orientation: 'horizontal',
-  isDisabled: false,
-  minValue: 0,
-  maxValue: 100,
-  step: 1
-};
-
-SliderCSS.argTypes = {
-  orientation: {
-    control: {
-      type: 'inline-radio',
-      options: ['horizontal', 'vertical']
-    }
-  }
-};
 
 export const TooltipExample = () => (
   <TooltipTrigger>
@@ -814,25 +744,6 @@ function MyMenuItem(props) {
         focused: isFocused,
         selected: isSelected
       })} />
-  );
-}
-
-function CustomThumb({index, children}) {
-  return (
-    <SliderThumb
-      index={index}
-      style={({isDragging, isFocusVisible}) => ({
-        width: 20,
-        height: 20,
-        borderRadius: '50%',
-        top: '50%',
-        // eslint-disable-next-line
-        backgroundColor: isFocusVisible ? 'orange' : isDragging
-          ? 'dimgrey'
-          : 'gray'
-      })}>
-      {children}
-    </SliderThumb>
   );
 }
